# ASGARproject/docs/swagger.yaml
basePath: /api
...
host: localhost:8080
info:
  contact:
    email: support@asgaravia.com
    name: API Support
    url: http://www.asgaravia.com/support
  description: API для системы заказа авиауслуг ASGAR Avia
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: ASGAR Avia API
  version: "1.0"
paths:
...
  /flight-servs:
    get:
      consumes:
      - application/json
      description: Возвращает список заявок текущего аутентифицированного пользователя
        (фильтрует "удален" и "черновик")
      parameters:
      - description: Фильтр по статусу
        in: query
        name: status
        type: string
      - description: 'Дата от (формат: YYYY-MM-DD)'
        in: query
        name: date_from
        type: string
      - description: 'Дата до (формат: YYYY-MM-DD)'
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Получить заявки пользователя
      tags:
      - flight-servs
  /flight-servs/{id}:
    delete:
      consumes:
      - application/json
      description: Помечает заявку как удаленную (меняет статус на 'удален')
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Удалить заявку
      tags:
      - flight-servs
    get:
      consumes:
      - application/json
      description: Возвращает детальную информацию о конкретной заявке
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Получить детали заявки
      tags:
      - flight-servs
    put:
      consumes:
      - application/json
      description: Обновляет данные существующей заявки
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: Обновляемые поля
        in: body
        name: updates
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Обновить заявку
      tags:
      - flight-servs
  /flight-servs/{id}/submit:
    put:
      consumes:
      - application/json
      description: Отправляет черновик заявки на обработку (меняет статус на 'формирование')
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Отправить заявку на обработку
      tags:
      - flight-servs
  /moderator/flight-servs:
    get:
      consumes:
      - application/json
      description: Возвращает список всех заявок в системе (только для модераторов)
      parameters:
      - description: Фильтр по статусу
        in: query
        name: status
        type: string
      - description: 'Дата от (формат: YYYY-MM-DD)'
        in: query
        name: date_from
        type: string
      - description: 'Дата до (формат: YYYY-MM-DD)'
        in: query
        name: date_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Получить все заявки (для модератора)
      tags:
      - flight-servs
  /moderator/flight-servs/{id}/complete:
    put:
      consumes:
      - application/json
      description: Завершает обработку заявки (только для модераторов)
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: 'Действие: complete (по умолчанию) или reject'
        in: query
        name: action
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Завершить заявку (для модератора)
      tags:
      - flight-servs
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"